refactor: extract createLangList
